:building_construction: Add exception handling. In detail, add exceptions and dto with enums.

+++ src/main/java/com/jellyfishmix/gouhaitakeaway/service/Impl/CommodityServiceImpl.java
@@ -1,10 +1,14 @@
 package com.jellyfishmix.gouhaitakeaway.service.Impl;
 
 import com.jellyfishmix.gouhaitakeaway.dao.CommodityDao;
+import com.jellyfishmix.gouhaitakeaway.dto.CommodityExecution;
 import com.jellyfishmix.gouhaitakeaway.entity.Commodity;
+import com.jellyfishmix.gouhaitakeaway.enums.CommodityStateEnum;
+import com.jellyfishmix.gouhaitakeaway.exceptions.CommodityOperationException;
 import com.jellyfishmix.gouhaitakeaway.service.CommodityService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
@@ -23,13 +27,27 @@ public class CommodityServiceImpl implements CommodityService {
     };
 
     /**
-     * 新增商品，返回值1为成功，-1为失败
+     * 添加商品
      * @param commodity
      * @return int
      */
     @Override
-    public int insertCommodity(Commodity commodity) {
-        return commodityDao.insertCommodity(commodity);
+    @Transactional
+    public CommodityExecution addCommodity(Commodity commodity) {
+        // 空值判断
+        if (commodity != null) {
+            try {
+                int effectedNum = commodityDao.insertCommodity(commodity);
+                if (effectedNum <= 0) {
+                    throw new CommodityOperationException("添加商品失败");
+                }
+            } catch (Exception e) {
+                throw new CommodityOperationException("添加商品失败，errMsg: " + e.toString());
+            }
+            return new CommodityExecution(CommodityStateEnum.SUCCESS);
+        } else {
+            return new CommodityExecution(CommodityStateEnum.EMPTY);
+        }
     }
 
     /**
@@ -38,8 +56,20 @@ public class CommodityServiceImpl implements CommodityService {
      * @return
      */
     @Override
-    public int updateCommodity(Commodity commodity) {
-        return commodityDao.updateCommodity(commodity);
+    public CommodityExecution updateCommodity(Commodity commodity) {
+        if (commodity != null) {
+            try {
+                int effectedNum = commodityDao.updateCommodity(commodity);
+                if (effectedNum <= 0) {
+                    throw new CommodityOperationException("修改商品失败");
+                }
+                return new CommodityExecution(CommodityStateEnum.SUCCESS);
+            } catch (Exception e) {
+                throw new CommodityOperationException("修改商品失败，errMsg: " + e.toString());
+            }
+        } else {
+            return new  CommodityExecution(CommodityStateEnum.EMPTY);
+        }
     }
 
     /**
@@ -48,7 +78,19 @@ public class CommodityServiceImpl implements CommodityService {
      * @return
      */
     @Override
-    public int deleteCommodity(Commodity commodity) {
-        return commodityDao.deleteCommodity(commodity);
+    public CommodityExecution deleteCommodity(Commodity commodity) {
+        if (commodity != null) {
+            try {
+                int effectedNum = commodityDao.deleteCommodity(commodity);
+                if (effectedNum <= 0) {
+                    throw new CommodityOperationException("删除商品失败");
+                }
+                return new CommodityExecution(CommodityStateEnum.SUCCESS);
+            } catch (Exception e) {
+                throw new CommodityOperationException("删除商品失败，errMsg: " + e.toString());
+            }
+        } else {
+            return new  CommodityExecution(CommodityStateEnum.EMPTY);
+        }
     }
 }
